rtld: Process caprelocs in _rtld_relocate_nonplt_self
This is conceptually where it belongs, as they're just another kind of
relocation that need processing before we can jump to normal C. This
avoids the weird passing of flags around and the need to have more
assembly.

Rather than using the start/stop symbols, we now use the dynamic tags
that have been present for a long time, akin to what is already done for
the rest of the relocating in _rtld_relocate_nonplt_self.

We also inline _rtld_do___caprelocs_self itself since it has MD parts
but is short.
